test(Message): 完善微信消息接收接口
1.完善微信消息接收接口

+++ src/main/java/com/ftc/weixintest/message/base/BaseMessage.java
@@ -31,7 +31,7 @@ public class BaseMessage {
      * 消息创建时间 （整型）
      */
     @XmlElement(name = "CreateTime")
-    private int createTime;
+    private long createTime;
 
     /**
      * 消息类型
